Add roadmap-renderer package
