--- lib/renderer/constants.ts
@@ -1,18 +1,0 @@
-export const FONT_SIZE = '13px';
-export const BORDER_WIDTH = 2.7;
-export const ARROW_WIDTH = 4;
-export const RECT_RADIUS = 2;
-
-export const DEFAULT_COLORS: Record<string, any> = {
-  black: ['#000'],
-  gray: ['#000', '#333', '#666', '#999', '#ccc', '#ddd', '#eee'],
-  white: ['#fff'],
-  red: ['#cf2a27', '#ea9999', '#eo6666', '#cc0000', '#990000', '#660000'],
-  orange: ['#ff9900', '#f9cb9c', '#f6b26b', '#e69138', '#b45f06', '#783f04'],
-  yellow: ['#ffff00', '#ffe599', '#ffd966', '#f1c232', '#bf9000', '#7f6000'],
-  green: ['#009e0f', '#b6d7a8', '#93c47d', '#6aa84f', '#38761d', '#274e13'],
-  cyan: ['#00ffff', '#a2c4c9', '#76a5af', '#45818e', '#134f5c', '#0c343d'],
-  blue: ['#2b78e4', '#9fc5f8', '#6fa8dc', '#597eaa', '#085394', '#073763'],
-  purple: ['#9900ff', '#b4a7d6', '#8e7cc3', '#674ea7', '#351c75', '#20124d'],
-  pink: ['#ff00ff', '#d5a6bd', '#c27ba0', '#a64d79', '#741b47', '#4c1130'],
-};
